pyproject.toml
[project]
name = "realmate-challenge"
version = "0.1.0"
description = ""
authors = [{ name = "Seu nome aqui", email = "seu.email@email.com.br" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["django (>=5.1.6,<6.0.0)", "djangorestframework (>=3.15.2,<4.0.0)", "django-rest-swagger (>=2.2.0,<3.0.0)", "setuptools (>=75.8.1,<76.0.0)", "drf-yasg (>=1.21.9,<2.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
